+++ flow-report/src/wrappers/report-renderer.tsx
@@ -6,9 +6,9 @@
 
 import {createContext, FunctionComponent} from 'preact';
 import {useContext, useMemo} from 'preact/hooks';
-import {CategoryRenderer} from '../../report/renderer/category-renderer';
-import {DetailsRenderer} from '../../report/renderer/details-renderer';
-import {DOM} from '../../report/renderer/dom';
+import {CategoryRenderer} from '../../../report/renderer/category-renderer';
+import {DetailsRenderer} from '../../../report/renderer/details-renderer';
+import {DOM} from '../../../report/renderer/dom';
 
 interface ReportRendererGlobals {
    dom: DOM,
